Fix for pickling problem
The issue was that Fit.__getattr__() was getting into an infinite loop
when unpickling a Fit object.  The problem was that when __getattr__
calls self.supported_distributions.keys(), it was triggering another
call to __getattr__.  Evidently, this happens when unpickling but not
otherwise because pickle.load() is building up the object, so more
attributes are missing then in general.  In any case, the fix is to
make a list of the keys that's outside the scope of the class, so that
__getattr__ doesn't end up calling itself.

For testing that the pickling and unpickling is working, I added
test_powerlaw_pickle.py, and walkobj.py.  The latter is used to
convert a Fit object to a nested list so that the unpickled object can
be compared to the original so as to confirm that the pickling is
working correctly.